Chat画面の更新をWebSocketに移行しました。
