+++ app/src/main/java/com/kyotob/client/chatList/ChatListActivity.kt
@@ -197,42 +197,3 @@ Java オブジェクトでキャメルケースに対応させるための設定
 
     }
 }
-
-// 非同期処理
-class DoAsync(private val handler: () -> Unit) : AsyncTask<Void, Void, Void>() {
-    override fun doInBackground(vararg params: Void?): Void? {
-        handler()
-        return null
-    }
-}
-// WebSocket
-@javax.websocket.ClientEndpoint
-class WebSocketEndPoint(private val handler: (msg: String) -> Unit) {
-
-    // Socket通信を開始するときに呼び出される
-    @OnOpen
-    fun onOpen(session: Session, config: EndpointConfig) {
-        println("client-[open] " + session)
-    }
-
-    // Message受信時に呼び出される
-    @OnMessage
-    fun onMessage(message: String, session: Session) {
-        println("client-[message][$message] $session")
-        if(message != "WebSocket通信を開始します。") { // 最初のメッセージは無視する
-            handler(message)
-        }
-    }
-
-    // Socket通信を終了するときに呼び出される
-    @OnClose
-    fun onClose(session: Session) {
-        println("client-[close] $session")
-    }
-
-    // ERRORのログを取る
-    @OnError
-    fun onError(session: Session?, t: Throwable?) {
-        println("client-[error] ${t?.message} $session")
-    }
-}
